Made the animator more general, this way it can be used for all entities/things that need animation instead of just the player

+++ src/assets/tile_sprites.h
@@ -4,8 +4,8 @@
 #define TILE_W 16
 #define TILE_H 16
 
-extern const int grass_sprite[TILE_H*TILE_W];
-extern const int stone_sprite[TILE_H*TILE_W];
-extern const int water_sprite[TILE_H*TILE_W];
+extern const short grass_sprite[TILE_H*TILE_W];
+extern const short stone_sprite[TILE_H*TILE_W];
+extern const short water_sprite[TILE_H*TILE_W];
 static const short dirt_sprite[TILE_W * TILE_H];
 #endif
